+++ apps/s3api/src/error.rs
@@ -4,15 +4,12 @@ pub type Result<T> = ::std::result::Result<T, Error>;
 
 #[derive(Fail, Debug)]
 pub enum Error {
-    #[fail(display = "subcommand missing")]
-    SubCommandMissing,
-
-    #[fail(display = "clap error")]
-    ClapError(clap::Error),
+    #[fail(display = "clap_task error")]
+    ClapTaskError(clap_task::Error),
 }
 
-impl From<clap::Error> for Error {
-    fn from(e: clap::Error) -> Self {
-        Error::ClapError(e)
+impl From<clap_task::Error> for Error {
+    fn from(e: clap_task::Error) -> Self {
+        Error::ClapTaskError(e)
     }
 }
