+++ apps/s3api/src/main.rs
@@ -1,14 +1,15 @@
 #[macro_use]
 extern crate clap;
+
 #[macro_use]
 extern crate failure;
 
 mod commands;
 mod error;
-use crate::commands::{Definition, Task};
-use crate::error::Error::SubCommandMissing;
-use clap::{App, ArgMatches};
 pub use error::Result as S3ApiResult;
+
+use clap::App;
+use clap_task::{Definition, TaskFinder};
 use std::process::exit;
 
 fn main() {
@@ -30,47 +31,13 @@ type ApiDef<'a, 'b> = Definition<'a, 'b, ApiResult>;
 
 fn run() -> ApiResult {
     let definitions: Vec<ApiDef> = vec![commands::put_object::create()];
-    let finder = TaskFinder::new(definitions)?;
+    let finder = TaskFinder::new(init(), definitions)?;
     let task = finder.require_task()?;
     task.run()
 }
 
-struct TaskFinder<'a, 'b, T> {
-    definitions: Vec<Definition<'a, 'b, T>>,
-    matches: ArgMatches<'a>,
-}
-
-impl<'a, 'b, T> TaskFinder<'a, 'b, T> {
-    fn new(definitions: Vec<Definition<'a, 'b, T>>) -> S3ApiResult<TaskFinder<'a, 'b, T>> {
-        let app = definitions
-            .iter()
-            .map(|task| task.define)
-            .fold(Self::init(), |acc, define| acc.subcommand(define()));
-
-        Ok(TaskFinder {
-            definitions,
-            matches: app.get_matches_safe()?,
-        })
-    }
-
-    fn init() -> App<'a, 'b> {
-        App::new(crate_name!())
-            .version(crate_version!())
-            .author(crate_authors!())
-    }
-
-    fn require_task(&'a self) -> S3ApiResult<Task<'a, 'b, T>> {
-        let to_task = |definition: &'a Definition<'a, 'b, T>| {
-            self.matches
-                .subcommand_matches(definition.name.to_string())
-                .map(|matches| Task {
-                    definition,
-                    matches,
-                })
-        };
-        self.definitions
-            .iter()
-            .find_map(to_task)
-            .ok_or(SubCommandMissing)
-    }
+fn init<'a, 'b>() -> App<'a, 'b> {
+    App::new(crate_name!())
+        .version(crate_version!())
+        .author(crate_authors!())
 }
