+++ libs/clap-task/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "clap-task"
+version = "0.0.1"
+edition = "2018"
+
+[dependencies]
+clap = "2.33"
+failure = "0.1"
